Added Controller, Repository, and Service for Book and Review (to be changed into ratings)
Book Class - contains the information for the books with the usable data
BookController Class - contains the functions getAllBooks() and getSingleBook. These will be used to create other methods.
BookRepository Interface - contains findBookByISBN method. This will be used to generate other methods later

The Review classes are similar to the Book ones